STVarExperiment

Documentation

type Atom = String

This module defines the type FlatTerm.

data VariableName

Constructors

VariableName Int String

data FlatTerm a

Constructors

Struct Atom [a]

Var VariableName

Wildcard

Cut Int

data Prolog

Constructors

P (Fix FlatTerm)

variableExtractor :: Fix FlatTerm -> [Fix FlatTerm]

variableNameExtractor :: Fix FlatTerm -> [VariableName]

variableSet :: [Fix FlatTerm] -> Set (Fix FlatTerm)

variableNameSet :: [VariableName] -> Set VariableName

varsToDictM :: (Ord a, Unifiable t) => Set a -> STBinding s (Map a (STVar s t))

uTermify :: Map VariableName (STVar s FlatTerm) -> UTerm FlatTerm (STVar s FlatTerm) -> UTerm FlatTerm (STVar s FlatTerm)

translateToUTerm :: Fix FlatTerm -> STBinding s (UTerm FlatTerm (STVar s FlatTerm), Map VariableName (STVar s FlatTerm))

vTermify :: Map Int VariableName -> UTerm FlatTerm (STVar s FlatTerm) -> UTerm FlatTerm (STVar s FlatTerm)

vTermify recursively converts UVar x into UTerm (VarA x). This is a subroutine of translateFromUTerm @. The resulting term has no (UVar x) subterms.

translateFromUTerm :: Map VariableName (STVar s FlatTerm) -> UTerm FlatTerm (STVar s FlatTerm) -> Prolog

unifyTerms :: Fix FlatTerm -> Fix FlatTerm -> Maybe (Map VariableName Prolog)

Unify two (E1 a) terms resulting in maybe a dictionary of variable bindings (to terms).

NB !!!! The current interface assumes that the variables in t1 and t2 are disjoint. This is likely a mistake that needs fixing

unifyTermsX :: Fix FlatTerm -> Fix FlatTerm -> ExceptT (UFailure FlatTerm (STVar s FlatTerm)) (STBinding s) (Map VariableName Prolog)

Unify two (E1 a) terms resulting in maybe a dictionary of variable bindings (to terms).

This routine works in the unification monad

mapWithKeyM :: (Ord k, Applicative m, Monad m) => (k -> a -> m b) -> Map k a -> m (Map k b)

makeDict :: Map VariableName (STVar s FlatTerm) -> STBinding s (Map VariableName Prolog)

makeDicts :: (Map VariableName (STVar s FlatTerm), Map VariableName (STVar s FlatTerm)) -> ExceptT (UFailure FlatTerm (STVar s FlatTerm)) (STBinding s) (Map VariableName Prolog)

recover the bindings for the variables of the two terms unified from the monad.

test1 :: ErrorT (UFailure FlatTerm (STVar s FlatTerm)) (STBinding s) (UTerm FlatTerm (STVar s FlatTerm), Map VariableName (STVar s FlatTerm))

test2 :: ErrorT (UFailure FlatTerm (STVar s FlatTerm)) (STBinding s) (UTerm FlatTerm (STVar s FlatTerm), Map VariableName (STVar s FlatTerm))

test3 :: ErrorT (UFailure FlatTerm (STVar s FlatTerm)) (STBinding s) (UTerm FlatTerm (STVar s FlatTerm), Map VariableName (STVar s FlatTerm))

test4 :: ErrorT (UFailure FlatTerm (STVar s FlatTerm)) (STBinding s) (UTerm FlatTerm (STVar s FlatTerm), Map VariableName (STVar s FlatTerm))

test5 :: ErrorT (UFailure FlatTerm (STVar s FlatTerm)) (STBinding s) (UTerm FlatTerm (STVar s FlatTerm), Map VariableName (STVar s FlatTerm))

test6 :: ErrorT (UFailure FlatTerm (STVar s FlatTerm)) (STBinding s) (UTerm FlatTerm (STVar s FlatTerm), Map VariableName (STVar s FlatTerm))

stExtract :: ErrorT (UFailure FlatTerm (STVar s FlatTerm)) (STBinding s) (UTerm FlatTerm (STVar s FlatTerm), Map VariableName (STVar s FlatTerm))

goTest :: Show b => (forall s. ErrorT (UFailure FlatTerm (STVar s FlatTerm)) (STBinding s) (UTerm FlatTerm (STVar s FlatTerm), Map VariableName (STVar s FlatTerm))) -> String

monadicUnification :: BindingMonad FlatTerm (STVar s FlatTerm) (STBinding s) => forall s. Fix FlatTerm -> Fix FlatTerm -> ErrorT (UFailure FlatTerm (STVar s FlatTerm)) (STBinding s) (UTerm FlatTerm (STVar s FlatTerm), Map VariableName (STVar s FlatTerm))

goUnify :: (forall s. BindingMonad FlatTerm (STVar s FlatTerm) (STBinding s) => ErrorT (UFailure FlatTerm (STVar s FlatTerm)) (STBinding s) (UTerm FlatTerm (STVar s FlatTerm), Map VariableName (STVar s FlatTerm))) -> [(VariableName, Prolog)]

f1 :: BindingMonad FlatTerm (STVar s FlatTerm) (STBinding s) => forall s. Map VariableName (STVar s FlatTerm) -> STBinding s [(VariableName, Prolog)]

fix1 :: Fix FlatTerm

fix2 :: Fix FlatTerm

fix3 :: Fix FlatTerm

fix4 :: Fix FlatTerm